fix: write a codemod to convert numbers to bigints
There are two types of integers supported in OpenAPI spec: `int32` and
`int64`. The current version of `orval`, however, converts all those
integers to js `number` and doesn't consider `int64` types as `bigint`.

To keep simplicity and handle `int64` types correctly, convert all of
numbers to bigints using a codemod. The codemod takes the result of
orval as input, converts the types and apply changes.

In addition, run the codemod in `generate-client` npm scripts, so the
future developers don't need to run two separate commands to get correct
client types.